feat: trail gradients, examples and refactor
- Split render module for separation of concerns
- Added fade_trail_color config option
- Improved hash distribution for randomized characters
- Updated ratatui to v0.29 for compatibility
- Added 5 examples
- README
