+++ .gitignore
@@ -12,6 +12,3 @@ target/
 # Generated by cargo mutants
 # Contains mutation testing data
 **/mutants.out*/
-
-# WIP files
-examples/
