+++ Cargo.toml
@@ -10,10 +10,10 @@ license = "GPL-3.0-or-later"
 readme = "README.md"
 keywords = ["cli", "utility", "tui", "simulation", "generator"]
 categories = ["command-line-utilities", "simulation", "science", "visualization", "mathematics" ]
-exclude = ["*.gif", "/.github"]
+exclude = ["*.gif", "/.github", "/examples", "/assets"]
 
 [dependencies]
-ratatui = "0.28"
+ratatui = "0.29"
 crossterm = "0.27"
 rand = "0.8"
 serde = { version = "1.0", features = ["derive"] }
