+++ examples/critter_highway.md
@@ -1,0 +1,24 @@
+# critter highway
+![critter highway](/assets/trmt_v0_3_0_critter_highway_example.png)
+
+### Config
+```toml
+[simulation]
+heads = 6
+rule = "WRSWNL"
+speed_ms = 10.0
+trail_length = 16
+color_cells = true
+seed = "nrfv6ndc"
+
+[display]
+colors = ["#667eea", "#764ba2", "#f093fb", "#f5576c"]
+fade_trail_color = ""
+state_based_colors = false
+live_colors = false
+randomize_heads = false
+randomize_trails = false
+head_char = ["██"]
+trail_char = ["▓▓"]
+cell_char = "░░"
+```
